requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
numpy==1.21.4
pandas==1.3.5
mysql-connector==2.2.9
